refactor(containers): move containers query to file
